Mask neutron services
Disable external ways to manage services (L3-agent/DHCP-agent/Metadata-agent/
ovs-agent) due to fact that these services are already managed by pacemaker.

Related-Bug: #1652748

+++ deployment/puppet/tweaks/manifests/ubuntu_service_override.pp
@@ -13,34 +13,55 @@
 #  prevented from being started as part of the installation process.
 #  Defaults to $name
 #
+# [*mask_service*]
+#  Boolean variable to mask service.
+#  Defaults to false
+
 define tweaks::ubuntu_service_override (
   $service_name = $name,
   $package_name = $name,
+  $mask_service = false,
 ) {
   if $::operatingsystem == 'Ubuntu' {
     if ! is_pkg_installed($package_name) {
-      # https://people.debian.org/~hmh/invokerc.d-policyrc.d-specification.txt
-      # use policy-rc.d to really ensure services don't get started on
-      # installation as service override files are only used if a job
-      # configuration file exists (see man 5 init)
-      $policyrc_file = '/usr/sbin/policy-rc.d'
+      # We need be able to mask services which should not be managed by manually
+      # to prevent situations like in https://bugs.launchpad.net/mos/+bug/1652748
+      if $::service_provider == 'systemd' {
+        if $mask_service {
+          ensure_resource('file', "/etc/systemd/system/${service_name}.service", {
+            ensure => 'link',
+            target => '/dev/null'
+          }
+        }
+      }
+      else {
+        ensure_resource('file', "/etc/systemd/system/${service_name}.service", {
+          ensure  => absent
+        })
+        # https://people.debian.org/~hmh/invokerc.d-policyrc.d-specification.txt
+        # use policy-rc.d to really ensure services don't get started on
+        # installation as service override files are only used if a job
+        # configuration file exists (see man 5 init)
+        $policyrc_file = '/usr/sbin/policy-rc.d'
 
-      # use ensure resource as we only want a single instance of the
-      # policy-rc.d file in the catalog
-      ensure_resource('file', 'create-policy-rc.d', {
-        ensure  => present,
-        path    => $policyrc_file,
-        content => "#!/bin/bash\nexit 101",
-        mode    => '0755',
-        owner   => 'root',
-        group   => 'root'
-      })
-      # use ensure resource as we only want a single remove exec in the catalog
-      ensure_resource('exec', 'remove-policy-rc.d', {
-        path    => [ '/sbin', '/bin', '/usr/bin', '/usr/sbin' ],
-        command => "rm -f ${policyrc_file}",
-        onlyif  => "test -f ${policyrc_file}",
-      })
+        # use ensure resource as we only want a single instance of the
+        # policy-rc.d file in the catalog
+        ensure_resource('file', 'create-policy-rc.d', {
+          ensure  => present,
+          path    => $policyrc_file,
+          content => "#!/bin/bash\nexit 101",
+          mode    => '0755',
+          owner   => 'root',
+          group   => 'root'
+        })
+        # use ensure resource as we only want a single remove exec in the catalog
+        ensure_resource('exec', 'remove-policy-rc.d', {
+          path    => [ '/sbin', '/bin', '/usr/bin', '/usr/sbin' ],
+          command => "rm -f ${policyrc_file}",
+          onlyif  => "test -f ${policyrc_file}",
+        })
+      }
+      endif
 
       File['create-policy-rc.d'] ->
         Package <| name == $package_name |> { provider =>  'apt_fuel' } ->
